requirements.txt
aioredis==1.2.0
asgiref==2.3.2
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
certifi==2019.3.9
channels==2.1.7
channels-redis==2.3.3
constantly==15.1.0
daphne==2.2.5
Django==2.1.7
django-cleanup==3.2.0
django-crispy-forms==1.7.2
gunicorn==19.9.0
hiredis==1.0.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
msgpack==0.6.1
olefile==0.46
Pillow==5.4.1
PyHamcrest==1.9.0
pytz==2018.9
six==1.12.0
Twisted==18.9.0
txaio==18.8.1
zope.interface==4.6.0
